:tada: actualizando refactor code
